fix(palette): preserve prior OAuth refresh token when /token omits it
